+++ app/src/main/java/com/pict/pbl/medswift/screens/symptoms/DiagnosisScreen.kt
@@ -1,6 +1,5 @@
 package com.pict.pbl.medswift.screens.symptoms
 
-import android.util.Log
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.fillMaxSize